Before (with GHC 9.10):
```
benchmarking Katip.Scribes.Handle/DevNull/full env, flush 1000 writes
time                 2.504 ms   (2.499 ms .. 2.509 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.514 ms   (2.510 ms .. 2.518 ms)
std dev              14.72 μs   (9.720 μs .. 24.45 μs)

benchmarking Katip.Scribes.Handle/TempFile/full env, flush 1000 writes
time                 4.831 ms   (4.798 ms .. 4.861 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 4.888 ms   (4.861 ms .. 4.954 ms)
std dev              121.2 μs   (44.10 μs .. 230.1 μs)
```

After:
```
benchmarking Katip.Scribes.Handle/DevNull/full env, flush 1000 writes
time                 2.164 ms   (2.104 ms .. 2.227 ms)
                     0.967 R²   (0.918 R² .. 0.993 R²)
mean                 2.507 ms   (2.391 ms .. 2.690 ms)
std dev              457.2 μs   (265.7 μs .. 788.3 μs)
variance introduced by outliers: 88% (severely inflated)

benchmarking Katip.Scribes.Handle/TempFile/full env, flush 1000 writes
time                 4.567 ms   (4.540 ms .. 4.599 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 4.617 ms   (4.597 ms .. 4.664 ms)
std dev              101.6 μs   (59.44 μs .. 168.7 μs)
```