@MichaelXavier 2024-08-27
Thanks so much @Bodigrim ! I've uploaded this to hackage as katip-0.8.8.2